fix pan / zoom listeners

+++ Calculator/src/main/java/com/android2/calculator3/Graph.java
@@ -45,7 +45,7 @@ public class Graph {
         });
         mGraphView.setZoomListener(new GraphView.ZoomListener() {
             @Override
-            public void zoomApplied(int level) {
+            public void zoomApplied(float level) {
                 mLogic.getGraphModule().updateGraph(Graph.this);
             }
         });
